requirements.txt
alembic
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.0
asyncpg==0.30.0
bcrypt==4.2.1
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.0
geojson==3.1.0
greenlet==3.1.1
h11==0.14.0
httptools==0.6.1
idna==3.10
ignr==2.2
jwt==1.3.1
Mako==1.3.8
MarkupSafe==3.0.2
numpy==2.0.2
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
pyproj==3.6.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.11
pytz==2024.2
PyYAML==6.0.2
rsa==4.9
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.0
uvloop==0.20.0
vulture==2.13
watchfiles==0.24.0
websockets==13.1

minio~=7.2.15
